# backend/openapi.yaml
openapi: 3.1.0
info:
  title: RAG Assistant Backend
  version: 0.2.1
paths:
  /api/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      description: 'Logout endpoint. Revokes the current JWT token by adding its JTI
        to the revocation list.

        The token will be invalid until its natural expiration.'
      operationId: logout_api_auth_logout_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/users:
    get:
      tags:
      - admin
      summary: List Users
      operationId: list_users_api_admin_users_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUserOut'
                title: Response List Users Api Admin Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/users/{user_id}/role:
    post:
      tags:
      - admin
      summary: Set User Role
      operationId: set_user_role_api_admin_users__user_id__role_post
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserRoleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/budgets/{user_id}:
    post:
      tags:
      - admin
      summary: Upsert Budget
      operationId: upsert_budget_api_admin_budgets__user_id__post
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetUpdate'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/audit:
    get:
      tags:
      - admin
      summary: List Audit Logs
      operationId: list_audit_logs_api_admin_audit_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminAuditEntry'
                title: Response List Audit Logs Api Admin Audit Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/system/summary:
    get:
      tags:
      - admin
      summary: System Summary
      operationId: system_summary_api_admin_system_summary_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSystemSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/watchdog/reset-stuck:
    post:
      tags:
      - admin
      summary: Admin Reset Stuck
      operationId: admin_reset_stuck_api_admin_watchdog_reset_stuck_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchdogResetRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchdogResetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/audit/deletions:
    get:
      tags:
      - admin
      summary: Get Deletion Audit
      operationId: get_deletion_audit_api_admin_audit_deletions_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeletionAuditEntry'
                title: Response Get Deletion Audit Api Admin Audit Deletions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AdminAuditEntry:
      properties:
        id:
          type: integer
          title: Id
        admin_user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Admin User Id
        action:
          type: string
          title: Action
        target_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Type
        target_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Id
        metadata_json:
          anyOf:
          - type: string
          - type: 'null'
          title: Metadata Json
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - action
      - created_at
      title: AdminAuditEntry
    AdminSystemSummary:
      properties:
        users:
          type: integer
          title: Users
        stores:
          type: integer
          title: Stores
        documents:
          type: integer
          title: Documents
      type: object
      required:
      - users
      - stores
      - documents
      title: AdminSystemSummary
    AdminUserOut:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          format: email
          title: Email
        is_admin:
          type: boolean
          title: Is Admin
        is_active:
          type: boolean
          title: Is Active
        admin_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Notes
        monthly_limit_usd:
          anyOf:
          - type: number
          - type: 'null'
          title: Monthly Limit Usd
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - email
      - is_admin
      - is_active
      - created_at
      title: AdminUserOut
    AdminUserRoleUpdate:
      properties:
        is_admin:
          type: boolean
          title: Is Admin
        admin_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Notes
      type: object
      required:
      - is_admin
      title: AdminUserRoleUpdate
...
    BudgetUpdate:
      properties:
        monthly_limit_usd:
          type: number
          maximum: 99999999.99
          minimum: 0.0
          title: Monthly Limit Usd
      type: object
      required:
      - monthly_limit_usd
      title: BudgetUpdate
...
    DeletionAuditEntry:
      properties:
        store_id:
          type: integer
          title: Store Id
        deleted_at:
          type: string
          format: date-time
          title: Deleted At
        deleted_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Deleted By
      type: object
      required:
      - store_id
      - deleted_at
      title: DeletionAuditEntry
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginIn:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginIn
...
    RegisterIn:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 72
          minLength: 6
          title: Password
      type: object
      required:
      - email
      - password
      title: RegisterIn
...
    TokenOut:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      title: TokenOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WatchdogResetRequest:
      properties:
        user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: User Id
        ttl_minutes:
          type: integer
          title: Ttl Minutes
          default: 30
      type: object
      title: WatchdogResetRequest
    WatchdogResetResponse:
      properties:
        reset_count:
          type: integer
          title: Reset Count
      type: object
      required:
      - reset_count
      title: WatchdogResetResponse